+++ NetFabric.Hyperlinq/Quantifier/All/All.ReadOnlySpan.cs
@@ -20,7 +20,7 @@ namespace NetFabric.Hyperlinq
         }
 
         [Pure]
-        public static bool All<TSource>(this ReadOnlySpan<TSource> source, Func<TSource, long, bool> predicate)
+        public static bool All<TSource>(this ReadOnlySpan<TSource> source, Func<TSource, int, bool> predicate)
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
